+++ bin/slim-cover.js
@@ -2,6 +2,7 @@
 
 const slimCover = require( "../src/shell" );
 const getArgs = require( "mri" );
+const path = require( "path" );
 
 const args = getArgs( process.argv.slice( 2 ) );
 
@@ -13,7 +14,7 @@ function sanitizeStringArg( arg = [] ) {
 	return arg;
 }
 
-const options = {
+let options = {
 	project: sanitizeStringArg( args.project ).pop(),
 	includes: sanitizeStringArg( args.include ),
 	excludes: sanitizeStringArg( args.exclude ),
@@ -22,5 +23,32 @@ const options = {
 		return { type, destination };
 	} )
 };
+
+const customConfigFile = sanitizeStringArg( args.config ).pop();
+const configFile = customConfigFile || ".slim-cover.json";
+
+let configFileOptions;
+
+try {
+	configFileOptions = require( path.resolve( configFile ) );
+}
+catch( e ) {
+	if( customConfigFile !== undefined ) {
+		console.error( "Error reading custom config file: ", e );
+	}
+}
+
+if( configFileOptions !== undefined ) {
+	function mergeOptionalArrays( left = [], right = [] ) {
+		return [ ...left, ...right ];
+	}
+	
+	options = {
+		project: mergeOptionalArrays( options.project, configFileOptions.project ).pop(),
+		includes: mergeOptionalArrays( options.includes, configFileOptions.includes ),
+		excludes: mergeOptionalArrays( options.excludes, configFileOptions.excludes ),
+		reporters: mergeOptionalArrays( options.reporters, configFileOptions.reporters )
+	};
+}
 
 slimCover( options );