+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Options may now be specified via config file.  See the updated readme for more details.
+
 ### Changed
 - File names in the output of both the terminal and codecov reporters will now be displayed as relative paths based on the project directory.
 - File names in the terminal reporter will now be formatted similarly to line numbers in order to stick out more.